+++ axelor-project/src/main/java/com/axelor/apps/project/module/ProjectModule.java
@@ -28,6 +28,8 @@ import com.axelor.apps.project.db.repo.TeamProjectRepository;
 import com.axelor.apps.project.db.repo.TeamTaskProjectRepository;
 import com.axelor.apps.project.service.ProjectService;
 import com.axelor.apps.project.service.ProjectServiceImpl;
+import com.axelor.apps.project.service.ProjectTemplateService;
+import com.axelor.apps.project.service.ProjectTemplateServiceImpl;
 import com.axelor.apps.project.service.TeamTaskProjectService;
 import com.axelor.apps.project.service.TeamTaskProjectServiceImpl;
 import com.axelor.apps.project.service.TimerTeamTaskService;
@@ -49,5 +51,6 @@ public class ProjectModule extends AxelorModule {
     bind(TeamRepository.class).to(TeamProjectRepository.class);
     bind(TimerTeamTaskService.class).to(TimerTeamTaskServiceImpl.class);
     bind(AbstractResourceBookingRepository.class).to(ResourceBookingRepository.class);
+    bind(ProjectTemplateService.class).to(ProjectTemplateServiceImpl.class);
   }
 }